+++ WPNinjas.PasswordGeneration.PowerShell/WPNinjas.PasswordGeneration/Functions/Invoke-PasswordGeneration.ps1
@@ -20,12 +20,16 @@ Function Invoke-PasswordGeneration(){
     .PARAMETER AllowedCharacters
         All characters in this string are allowed as characters during password generation. If no characters are defined the following default set will be used "ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz.:;,-_!?$%*=+&<>@#()23456789".
 
+    .PARAMETER CheckComplexity
+        If set the password will regenerated until it contains characters from at least 3 groups out of 4 (Uper/Lower/Number/SpecialChar)
+
     .NOTES
     Author: Thomas Kurth/baseVISION
-    Date:   13.6.2021
+    Date:   10.06.2022
 
     History
         See Release Notes in Github.
+        10.6.2022: Add COmplexity Check
 
     #>
     [CmdletBinding()]
@@ -34,8 +38,10 @@ Function Invoke-PasswordGeneration(){
         [Parameter(Mandatory=$false)]
         [int]$Length = 16,
         [Parameter(Mandatory=$false)]
-        [String]$AllowedCharacters = "ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz.:;,-_!?$%*=+&<>@#()23456789"
-    )
+        [String]$AllowedCharacters = "ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz.:;,-_!?$%*=+&<>@#()23456789",
+        [Switch]$RequireComplexity
+    
+        )
       
 
     #region Initialization
@@ -45,9 +51,11 @@ Function Invoke-PasswordGeneration(){
 
     #region Main Script
     ########################################################
-    
-    [WPNinjas.PasswordGeneration.PwService]::GetRandomPassword($Length,$AllowedCharacters)
-
+    if($RequireComplexity){
+        [WPNinjas.PasswordGeneration.PwService]::GetRandomComplexPassword($Length,$AllowedCharacters)
+    } else {
+        [WPNinjas.PasswordGeneration.PwService]::GetRandomPassword($Length,$AllowedCharacters)
+    }
     #endregion
     #region Finishing
     ########################################################