+++ WPNinjas.PasswordGeneration.PowerShell/WPNinjas.PasswordGeneration/Functions/Invoke-SecurePasswordGeneration.ps1
@@ -20,6 +20,9 @@ Function Invoke-SecurePasswordGeneration(){
     .PARAMETER AllowedCharacters
         All characters in this string are allowed as characters during password generation. If no characters are defined the following default set will be used "ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz.:;,-_!?$%*=+&<>@#()23456789".
 
+    .PARAMETER CheckComplexity
+        If set the password will regenerated until it contains characters from at least 3 groups out of 4 (Uper/Lower/Number/SpecialChar)
+
     .NOTES
     Author: Thomas Kurth/baseVISION
     Date:   13.6.2021
@@ -34,7 +37,8 @@ Function Invoke-SecurePasswordGeneration(){
         [Parameter(Mandatory=$false)]
         [int]$Length = 16,
         [Parameter(Mandatory=$false)]
-        [String]$AllowedCharacters = "ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz.:;,-_!?$%*=+&<>@#()23456789"
+        [String]$AllowedCharacters = "ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz.:;,-_!?$%*=+&<>@#()23456789",
+        [Switch]$RequireComplexity
     )
       
 
@@ -46,7 +50,11 @@ Function Invoke-SecurePasswordGeneration(){
     #region Main Script
     ########################################################
     
-    [WPNinjas.PasswordGeneration.PwService]::GetRandomPasswordSecure($Length,$AllowedCharacters)
+    if($RequireComplexity){
+        [WPNinjas.PasswordGeneration.PwService]::GetRandomPasswordSecure($Length,$AllowedCharacters)
+    } else {
+        [WPNinjas.PasswordGeneration.PwService]::GetRandomComplexPasswordSecure($Length,$AllowedCharacters)
+    }
 
     #endregion
     #region Finishing