+++ WPNinjas.PasswordGeneration.PowerShell/ReleaseNotes.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## 1.0.1 - 10.6.2022 - Thomas Kurth
+
+ - Add Complexity Check which tries to generate a complex password up to 10 times.
+
 ## 1.0.0 - 13.06.2021 - Thomas Kurth
 
 - Initial Version